+++ src/app/services/sdk/presenters/index.tsx
@@ -9,7 +9,7 @@
 
 import {C_MSG} from '../const';
 import {AuthSentCode, AuthCheckedPhone, AuthAuthorization, AuthRecalled} from '../messages/chat.api.auth_pb';
-import {ContactsImported, ContactsMany} from '../messages/chat.api.contacts_pb';
+import {BlockedContactsMany, ContactsImported, ContactsMany} from '../messages/chat.api.contacts_pb';
 import {Bool, Error, Label, LabelsMany, MessageContainer} from '../messages/chat.core.types_pb';
 import {MessagesDialogs, MessagesMany, MessagesSent} from '../messages/chat.api.messages_pb';
 import {Dialog, Group, GroupFull, PeerNotifySettings, User} from '../messages/chat.core.types_pb';
@@ -85,6 +85,8 @@ export default class Presenter {
                 return Label.deserializeBinary(data);
             case C_MSG.AccountPassword:
                 return AccountPassword.deserializeBinary(data);
+            case C_MSG.BlockedContactsMany:
+                return BlockedContactsMany.deserializeBinary(data);
             default:
                 return null;
         }